simplify lib to one script
